# swagger.yaml
openapi: 3.0.0
info:
  title: E-Learning Platform API
  version: 1.0.0
  description: |
    توثيق جميع الـ endpoints الخاصة بمنصة التعليم الإلكتروني (Node.js/Express/MongoDB)
    
    **جميع المسارات المحمية تتطلب JWT في الهيدر Authorization: Bearer TOKEN**
servers:
  - url: http://localhost:5000
paths:
...
  /user:
    get:
      summary: جلب كل المستخدمين (أدمن فقط)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: قائمة المستخدمين
...
  /user/{id}:
    put:
      summary: تحديث بيانات مستخدم (أدمن أو المستخدم نفسه)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: تم التحديث
    delete:
      summary: حذف مستخدم (أدمن فقط)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: تم الحذف
...
  /admin/users:
    get:
      summary: جلب كل المستخدمين (أدمن فقط)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: قائمة المستخدمين
  /admin/users/{id}:
    put:
      summary: تحديث بيانات مستخدم (أدمن فقط)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: تم التحديث
    delete:
      summary: حذف مستخدم (أدمن فقط)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: تم الحذف
  /admin/create-teacher:
    post:
      summary: إنشاء حساب معلم (أدمن فقط)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                phone:
                  type: string
                academicYear:
                  type: integer
                profileImage:
                  type: string
                  format: binary
      responses:
        '201':
          description: تم إنشاء المعلم (مع أو بدون صورة)
  /admin/approve-carnet:
    post:
      summary: اعتماد الكارنيه (أدمن فقط)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: تم الاعتماد
  /admin/reject-carnet:
    post:
      summary: رفض الكارنيه (أدمن فقط)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: تم الرفض
  /admin/stats:
    get:
      summary: جلب إحصائيات عامة للوحة تحكم الأدمن
      security:
        - bearerAuth: []
      responses:
        '200':
          description: إحصائيات

  # COURSES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT